-------------------------------------------------------------------------------
    Test name                                    | head[ms] | base[ms] |  ratio   |
    -------------------------------------------------------------------------------
    groupby_multi_index                          | 1008.6970 | 1861.1110 |   0.5420 |
    -------------------------------------------------------------------------------
    Test name                                    | head[ms] | base[ms] |  ratio   |
    -------------------------------------------------------------------------------

    Ratio < 1.0 means the target commit is faster then the baseline.
    Seed used: 1234

    Target [907e1b7] : performance improvements in multi-key groupby
    Base   [5dc8009] : Merge pull request #9418 from sinhrks/rn_20150205

    DOC: Fix release note format