@jreback 2015-02-06
pls add a release note in perf section, otherwise looks good.