@behzadnouri 2015-02-07
@jreback added the release note