@jreback 2015-02-07
thanks!

we should prob rerun these benchmarks to see how much better things are now

https://github.com/Rdatatable/data.table/wiki/Benchmarks-%3A-Grouping